requirements.txt
azure-ai-documentintelligence
azure-core
pandas
python-dotenv
pydantic
openai
chromadb
langgraph
